requirements.txt
requests==2.22.0
#bs4==0.0.1
scrapy==1.8.0
html2text-2019.9.26
boto3=1.11.0